+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Improvements
 - Added `IceCandidateStats` to `StatsReport` which gives insight into individual ice candidates
 - Removed forced ice-restart when synced responses are processed
 - Improved the detection of, and recovery from media connection failures. Worst case recovery times have been reduced by up to 10 seconds.
+- Added handle to native audio track. This will assist developers when using advanced features like audio track sink.
 
 Known issues